JOINTS 2020/final/web/OldSkool/old_skool/requirements.txt
bcrypt
flask
flask_generic_views
flask-sqlalchemy
passlib
wtforms
flask_login
multidict
